Improve eternals' time of day implementation (Present)
Special case indefinite unit types to not use the usual bridge conversion code.
Bridges were originally used for managing the conversion between a definite point
in time and months. The introduction of eternals adds an additional kind of unit:
the indefinite unit. These units must be handled specially as there is no remainder
component that must be dealt with.